mysite/requirementspdf.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.5
django-filter==23.2
djangorestframework==3.14.0
fpdf==1.7.2
idna==3.4
importlib-metadata==6.6.0
Markdown==3.4.3
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
python-barcode==0.14.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
zipp==3.15.0
